1:添加时间转换函数示例 2:显示进度示例 3:文件拷贝示例 4:文件权健设置函数示例

+++ mktime_localtime.c
@@ -23,11 +23,8 @@ int main ()
    else
    {
       strftime(buffer, sizeof(buffer), "%c", &info );
-      printf("%s   %d\n",buffer,  ret);
+      printf("%s   %d\n",buffer,  (int)ret);
    }
-
-
-
 
    struct tm *tm_info;
    time_t rawtime = ret;
